feat: add click integer list parameter type
